# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flash Net Network Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface LoginRequest {
  phone: string;
  password: string;
}

export type UserProfileRole =
  (typeof UserProfileRole)[keyof typeof UserProfileRole];

export const UserProfileRole = {
  owner: "owner",
  finance_manager: "finance_manager",
  supervisor: "supervisor",
  tech_engineer: "tech_engineer",
} as const;

export interface UserProfile {
  id: number;
  name: string;
  phone: string;
  role: UserProfileRole;
  isActive: boolean;
  createdAt: string;
}

export interface LoginResponse {
  token: string;
  user: UserProfile;
}

export type CreateUserRequestRole =
  (typeof CreateUserRequestRole)[keyof typeof CreateUserRequestRole];

export const CreateUserRequestRole = {
  finance_manager: "finance_manager",
  supervisor: "supervisor",
  tech_engineer: "tech_engineer",
} as const;

export interface CreateUserRequest {
  name: string;
  phone: string;
  password: string;
  role: CreateUserRequestRole;
}

export type UpdateUserRequestRole =
  (typeof UpdateUserRequestRole)[keyof typeof UpdateUserRequestRole];

export const UpdateUserRequestRole = {
  finance_manager: "finance_manager",
  supervisor: "supervisor",
  tech_engineer: "tech_engineer",
} as const;

export interface UpdateUserRequest {
  name?: string;
  phone?: string;
  role?: UpdateUserRequestRole;
}

export type HotspotPointStatus =
  (typeof HotspotPointStatus)[keyof typeof HotspotPointStatus];

export const HotspotPointStatus = {
  active: "active",
  active_incomplete: "active_incomplete",
  ready: "ready",
  empty: "empty",
  stopped: "stopped",
} as const;

export interface HotspotPoint {
  id: number;
  name: string;
  location: string;
  status: HotspotPointStatus;
  supervisorId?: number | null;
  notes?: string | null;
  createdAt: string;
}

export type BroadbandPointStatus =
  (typeof BroadbandPointStatus)[keyof typeof BroadbandPointStatus];

export const BroadbandPointStatus = {
  active: "active",
  active_incomplete: "active_incomplete",
  ready: "ready",
  empty: "empty",
  stopped: "stopped",
} as const;

export interface BroadbandPoint {
  id: number;
  name: string;
  location: string;
  status: BroadbandPointStatus;
  supervisorId?: number | null;
  speed?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface SalesPoint {
  id: number;
  name: string;
  location: string;
  managerId?: number | null;
  notes?: string | null;
  createdAt: string;
}