# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flash Net Network Management API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with phone and password
 */
export const LoginBody = zod.object({
  phone: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    phone: zod.string(),
    role: zod.enum(["owner", "finance_manager", "supervisor", "tech_engineer"]),
    isActive: zod.boolean(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  role: zod.enum(["owner", "finance_manager", "supervisor", "tech_engineer"]),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all users
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  role: zod.enum(["owner", "finance_manager", "supervisor", "tech_engineer"]),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Create a new user (Owner only)
 */
export const CreateUserBody = zod.object({
  name: zod.string(),
  phone: zod.string(),
  password: zod.string(),
  role: zod.enum(["finance_manager", "supervisor", "tech_engineer"]),
});

/**
 * @summary Get user by ID
 */
export const GetUserParams = zod.object({
  id: zod.coerce.number(),
});

export const GetUserResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  role: zod.enum(["owner", "finance_manager", "supervisor", "tech_engineer"]),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Update a user (Owner only)
 */
export const UpdateUserParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateUserBody = zod.object({
  name: zod.string().optional(),
  phone: zod.string().optional(),
  role: zod.enum(["finance_manager", "supervisor", "tech_engineer"]).optional(),
});

export const UpdateUserResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  role: zod.enum(["owner", "finance_manager", "supervisor", "tech_engineer"]),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Activate or deactivate a user (Owner only)
 */
export const ToggleUserActiveParams = zod.object({
  id: zod.coerce.number(),
});

export const ToggleUserActiveResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  phone: zod.string(),
  role: zod.enum(["owner", "finance_manager", "supervisor", "tech_engineer"]),
  isActive: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List hotspot points
 */
export const ListHotspotPointsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  status: zod.enum([
    "active",
    "active_incomplete",
    "ready",
    "empty",
    "stopped",
  ]),
  supervisorId: zod.number().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListHotspotPointsResponse = zod.array(
  ListHotspotPointsResponseItem,
);

/**
 * @summary List broadband points
 */
export const ListBroadbandPointsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  status: zod.enum([
    "active",
    "active_incomplete",
    "ready",
    "empty",
    "stopped",
  ]),
  supervisorId: zod.number().nullish(),
  speed: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListBroadbandPointsResponse = zod.array(
  ListBroadbandPointsResponseItem,
);

/**
 * @summary List sales points
 */
export const ListSalesPointsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  location: zod.string(),
  managerId: zod.number().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
});
export const ListSalesPointsResponse = zod.array(ListSalesPointsResponseItem);