# lib/api-zod/src/generated/types/broadbandPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flash Net Network Management API
 * OpenAPI spec version: 0.1.0
 */
import type { BroadbandPointStatus } from "./broadbandPointStatus";

export interface BroadbandPoint {
  id: number;
  name: string;
  location: string;
  status: BroadbandPointStatus;
  supervisorId?: number | null;
  speed?: string | null;
  notes?: string | null;
  createdAt: Date;
}