# lib/api-zod/src/generated/types/hotspotPoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flash Net Network Management API
 * OpenAPI spec version: 0.1.0
 */
import type { HotspotPointStatus } from "./hotspotPointStatus";

export interface HotspotPoint {
  id: number;
  name: string;
  location: string;
  status: HotspotPointStatus;
  supervisorId?: number | null;
  notes?: string | null;
  createdAt: Date;
}