# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Flash Net Network Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileRole } from "./userProfileRole";

export interface UserProfile {
  id: number;
  name: string;
  phone: string;
  role: UserProfileRole;
  isActive: boolean;
  createdAt: Date;
}